# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.3"
  title: titulo
  description: descrição

# during dev, should point to your local machine
host: localhost:8000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
 Bearer:
  description: |
   Autenticação da API se baseia nos seguintes recursos:

   - API Key
   - OAuth2

   O acesso a API ncecessita de um token JWT válido que deve ser passado no cabeçalho no campo 'Authorization'.

   Um token JWT válido é gerado pela API e retornado na resposta a rota
   /login uma vez que possua um email & senha válido.

   A seguinte sintaxe deve ser utilizada no cabeçalho das requisições 'Authorization' :
       Bearer: xxxxxx.yyyyyyy.zzzzzz
  type: apiKey
  name: Authorization
  in: header
paths:
...
  /api/user/perfil:
    x-swagger-router-controller: user
    get:
      description: Retorna o perfil do usuario logado
      summary: Lista de Usuários do MGI
      operationId: getPerfil
      tags:
        - Infraestrutura
      security:
        - Bearer: []
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/GetUserPerfilResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/indicador/importacao:
    x-swagger-router-controller: indicador
    get:
      description: Retorna os dados de importação dos indicadores cadastrados no sistema (3 - Importação e 4 - Formulário)
      summary: Lista de indicadores que podem ser importados
      operationId: getIndicadoresImportacao
      tags:
        - Indicadores
        - Importação
      parameters:
        - name: tipo
          in: query
          required: true
          enum: [3, 4]
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/GetIndicadorConsulta"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador:
    x-swagger-router-controller: indicador
    get:
      description: Retorna os indicadores cadastrados no sistema
      summary: Lista de indicadores
      operationId: getIndicadores
      tags:
        - Indicadores
      parameters:
        - name: limit
          in: query
          required: false
          type: number
        - name: offset
          in: query
          required: false
          type: number
        - name: query
          in: query
          required: false
          type: string
        - name: secretaria
          in: query
          required: false
          type: number

        - name: tag
          in: query
          required: false
          type: array
          items:
            type: number
        - name: codigos
          in: query
          required: false
          type: array
          items:
            type: string

      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/GetIndicadorConsulta"
        default:
          description: Erro Tratado
          schema:
            $ref: "#/definitions/ErrorResponseTratado"
    post:
      description: Cria um novo indicador
      summary: Cria um novo indicador
      operationId: createIndicador
      tags:
        - Indicadores
      parameters:
        - name: indicador
          in: body
          required: true
          schema:
            $ref: '#/definitions/IndicadorAtualizacao'
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro Tratado
          schema:
            $ref: "#/definitions/ErrorResponseTratado"
...
  /api/indicador-id/{id}:
    x-swagger-router-controller: indicador
    get:
      description: Le um registro de indicador
      summary: Retorna um registro de um indicador
      operationId: getIndicadorPorId
      tags:
        - Indicadores
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Sucesso
          schema:
            type: object
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{codigo}:
    x-swagger-router-controller: indicador
    get:
      description: Le um registro de indicador
      summary: Retorna um registro de um indicador
      operationId: getIndicador
      tags:
        - Indicadores
      parameters:
        - name: codigo
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Sucesso
          schema:
            type: object
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Apaga um registro de indicador
      summary: Apaga um registro de indicador
      operationId: deleteIndicador
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: codigo
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseTratado"
    put:
      description: Atualiza as informações de um registro de indicador
      summary: Atualiza as informações de um indicador
      operationId: editaIndicador
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: codigo
          type: string
          in: path
          required: true
        - name: indicador
          in: body
          required: true
          schema:
            $ref: '#/definitions/IndicadorAtualizacao'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{codigo}/conceituacao:
    x-swagger-router-controller: indicador
    patch:
      description: Atualiza conceituação de um indicador
      summary: Atualiza uma conceituação de um indicador
      operationId: updateConceituacao
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: codigo
          type: string
          in: path
          required: true
        - name: conceituacao
          in: body
          required: true
          schema:
            $ref: '#/definitions/Conceituacao'
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{id_pai}/indicador-relacionado/{id}:
    x-swagger-router-controller: indicador
    patch:
      description: Adiciona relação a um indicador
      summary: Adiciona a relação entre indicadores
      operationId: addIndicadorRelacionado
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: id_pai
          type: number
          in: path
          required: true
        - name: id
          in: path
          required: true
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Rompe a relação com um indicador
      summary: Rompe a relação entre indicadores
      operationId: deleteIndicadorRelacionado
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: id_pai
          type: number
          in: path
          required: true
        - name: id
          in: path
          required: true
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{id}/categoria-analise/{categoria_analise}:
    x-swagger-router-controller: indicador
    patch:
      description: Adiciona categoria relacionada a um indicador
      summary: Adiciona uma categoria de análise a um indicador
      operationId: addCategoriaAnalise
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: categoria_analise
          type: string
          in: path
          required: true
        - name: id
          in: path
          required: true
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Rompe a relação com um indicador
      summary: Rompe a relação entre indicadores
      operationId: deleteCategoriaAnalise
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: categoria_analise
          type: string
          in: path
          required: true
        - name: id
          in: path
          required: true
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{id}/responsavel-gerencial/{responsavel_gerencial}:
    x-swagger-router-controller: indicador
    patch:
      description: Adiciona responsavel gerencial relacionado a um indicador
      summary: Adiciona um responsável geárencial
      operationId: addResponsavelGerencial
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: responsavel_gerencial
          type: string
          in: path
          required: true
        - name: id
          in: path
          required: true
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Rompe a relação entre o responsável gerencial e um indicador
      summary: Rompe a relação entre indicadores e responsável gerencial
      operationId: deleteResponsavelGerencial
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: responsavel_gerencial
          type: string
          in: path
          required: true
        - name: id
          in: path
          required: true
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{id}/responsavel-tecnico/{responsavel_tecnico}:
    x-swagger-router-controller: indicador
    patch:
      description: Adiciona responsável tniéco relacionado a um indicador
      summary: Adiciona uma responsável técnico a um indicador
      operationId: addResponsavelTecnico
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: responsavel_tecnico
          type: string
          in: path
          required: true
        - name: id
          in: path
          required: true
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Rompe a relação de responsabilidade técnica com um indicador
      summary: Rompe a relação entre responsável técnico e um indicador
      operationId: deleteResponsavelTecnico
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: responsavel_tecnico
          type: string
          in: path
          required: true
        - name: id
          in: path
          required: true
          type: number
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{codigo}/interpretacao:
    x-swagger-router-controller: indicador
    patch:
      description: Atualiza a interpretação de um indicador
      summary: Atualiza o campo interpretação de um indicador
      operationId: updateInterpretacao
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: codigo
          type: string
          in: path
          required: true
        - name: interpretacao
          in: body
          required: true
          schema:
            $ref: '#/definitions/Interpretacao'
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{codigo}/procedimento_operacional:
    x-swagger-router-controller: indicador
    patch:
      description: Atualiza o procedimento operacional de um indicador
      summary: Atualiza o procedimento operacional de um indicador
      operationId: updateProcedimentoOperacional
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: codigo
          type: string
          in: path
          required: true
        - name: procedimento_operacional
          in: body
          required: true
          schema:
            $ref: '#/definitions/ProcedimentoOperacional'
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/indicador/{codigo}/uso:
    x-swagger-router-controller: indicador
    patch:
      description: Atualiza a interpretação de um indicador
      summary: Atualiza a interpretação de um indicador
      operationId: updateUso
      tags:
        - Indicadores
      security:
        - Bearer: []
      parameters:
        - name: codigo
          type: string
          in: path
          required: true
        - name: interpretacao
          in: body
          required: true
          schema:
            $ref: '#/definitions/Uso'
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Conceituacao:
    type: object
    properties:
      conceituacao:
        type: string
  Interpretacao:
    type: object
    properties:
      interpretacao:
        type: string
  ProcedimentoOperacional:
    type: object
    properties:
      procedimento_operacional:
        type: string
  Uso:
    type: object
    properties:
      usos:
        type: string
...
  GetUserPerfilResponse:
    required:
      - perfil
    properties:
      perfil:
        type: object
...
  GetIndicadorConsulta:
    required:
      - count
      - rows
    properties:
      count:
        type: integer
      rows:
        type: array
        items:
          type: object
  SuccessResponse:
    required:
      - codret
      - mensagem
    properties:
      codret:
        type: number
      mensagem:
        type: string
  ErrorResponseTratado:
    description: Erro generico
    type: object
  ErrorResponse:
    required:
      - message
    properties:
      codret:
        type: string
      message:
        type: string